+++ Fitbit.Portable/Models/DatasetInterval.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Fitbit.Api.Portable.Models
+{
+    public class DatasetInterval
+    {
+        public DateTime Time { get; set; }
+        public int Value { get; set; }
+    }
+}
